Finalize code cleaning and static implicit / explicit operators use
