+++ GB.NetApi.Infrastructure.Database/Repositories/AuthenticateUserRepository.cs
@@ -2,11 +2,9 @@
 using GB.NetApi.Domain.Models.Interfaces.Repositories;
 using GB.NetApi.Infrastructure.Database.Contexts;
 using GB.NetApi.Infrastructure.Database.DAOs.Identity;
-using GB.NetApi.Infrastructure.Database.Extensions;
 using GB.NetApi.Infrastructure.Database.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using System;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace GB.NetApi.Infrastructure.Database.Repositories
@@ -26,9 +24,8 @@ namespace GB.NetApi.Infrastructure.Database.Repositories
             using (var context = Repository.InstanciateContext())
             {
                 Task<UserDao> function() => GetAsync(context, userEmail);
-                var result = await Repository.ExecuteAsync(function).ConfigureAwait(false);
-
-                return Transform(result);
+                
+                return (AuthenticateUser)await Repository.ExecuteAsync(function).ConfigureAwait(false);
             }
         }
 
@@ -51,23 +48,6 @@ namespace GB.NetApi.Infrastructure.Database.Repositories
             .ThenInclude(e => e.Resource)
             .SingleOrDefaultAsync(e => e.Email == userEmail)
             .ConfigureAwait(false);
-
-        private static AuthenticateUser Transform(UserDao user)
-        {
-            if (user is null)
-                return default;
-
-            var claims = user.UserRoles.ToRoleClaims().Union(user.UserClaims.ToClaims());
-            var permissions = user.UserRoles.GetPermissionNames();
-
-            return new AuthenticateUser()
-            {
-                ID = user.Id,
-                Name = user.UserName,
-                Claims = claims,
-                PermissionNames = permissions
-            };
-        }
 
         #endregion
     }
